requirements.txt
# Core ML/NLP Libraries
sentence-transformers>=2.2.2
transformers>=4.30.0
faiss-cpu>=1.7.4
numpy>=1.21.0

# Web Scraping & HTTP
requests>=2.28.0
beautifulsoup4>=4.11.0

# LLM Clients
groq>=0.4.1

# Data Validation
pydantic>=1.10.0

# Environment & Configuration
python-dotenv>=0.19.0

# Graph Processing
networkx>=2.8.0

# Optional: For enhanced web scraping
# selenium>=4.0.0
# playwright>=1.25.0

# Development & Testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
